refactor: move floating element logic
