+++ packages/milkup-equations/src/inline/InlineEquationEditor.tsx
@@ -5,62 +5,7 @@ import React, { ChangeEvent, Ref, RefObject, useEffect } from "react";
 import { InlineMath } from "react-katex";
 
 import "./InlineEquationEditor.css";
-
-const VERTICAL_GAP = 10;
-const HORIZONTAL_OFFSET = 5;
-type Position = "auto" | "above" | "below";
-
-export function setFloatingElemPosition(
-  targetRect: DOMRect | null,
-  floatingElem: HTMLElement,
-  anchorElem: HTMLElement,
-  isLink: boolean = false,
-  position: Position = "auto",
-  centering: boolean = false,
-  verticalGap: number = VERTICAL_GAP,
-  horizontalOffset: number = HORIZONTAL_OFFSET,
-): void {
-  const scrollerElem = anchorElem.parentElement;
-
-  if (targetRect === null || !scrollerElem) {
-    floatingElem.style.opacity = "0";
-    floatingElem.style.transform = "translate(-10000px, -10000px)";
-    return;
-  }
-
-  const floatingElemRect = floatingElem.getBoundingClientRect();
-  const anchorElementRect = anchorElem.getBoundingClientRect();
-  const editorScrollerRect = scrollerElem.getBoundingClientRect();
-
-  let top =
-    targetRect.top +
-    floatingElemRect.height * (position == "below" ? 1 : -1) -
-    verticalGap;
-  let left = targetRect.left - horizontalOffset;
-
-  if (top < editorScrollerRect.top) {
-    // adjusted height for link element if the element is at top
-    top +=
-      floatingElemRect.height +
-      targetRect.height +
-      verticalGap * (isLink ? 9 : 2);
-  }
-
-  if (left + floatingElemRect.width > editorScrollerRect.right) {
-    left = editorScrollerRect.right - floatingElemRect.width - horizontalOffset;
-  }
-
-  top -= anchorElementRect.top;
-  left -= anchorElementRect.left;
-
-  if (centering) {
-    left +=
-      (targetRect.right - targetRect.left) / 2 - floatingElemRect.width / 2;
-  }
-
-  floatingElem.style.opacity = "1";
-  floatingElem.style.transform = `translate(${left}px, ${top}px)`;
-}
+import { setFloatingElemPosition } from "./utils/setFloatingElemPosition";
 
 type EquationEditorProps = {
   editor: LexicalEditor;
